⚠️Refreashed project configurations
Created a new project,
migrated previous files to new project.

Reason being previous project is prone to 'Connection String Converter' exception. Most likely becuase of models..? Since I've found that in Web.config line 27, the connection metadata seems to be referring to models, and our models here is just to make ourselves understand the database easier without affeting the current project.

Online resolutions are to create a new project. And here we are, in our awkward self. Good as new.